optional layer refactor
