+++ elements/layercontrol/src/helpers/filter-layers.js
@@ -1,0 +1,32 @@
+/**
+ * Filter all map layers by property
+ *
+ * @param {Array<import("ol/layer").Layer>} layers
+ * @param {string} key
+ * @param {any} value
+ * @returns {Array<import("ol/layer").Layer>} found layers
+ */
+export default function filterLayers(layers, key, value) {
+  /**
+   * @type {any[]}
+   */
+  let found = [];
+  /**
+   *
+   * @param {any[]} searchLayers
+   * @param {string} key
+   * @param {any} value
+   */
+  const search = (searchLayers, key, value) => {
+    found = [...found, ...searchLayers.filter((l) => l.get(key) === value)];
+    const groups = searchLayers.filter((l) => l.getLayers);
+    if (groups.length > 0) {
+      groups.forEach((group) =>
+        search(group.getLayers().getArray(), key, value)
+      );
+    }
+    return found;
+  };
+  search(layers, key, value);
+  return found;
+}
