+++ elements/layercontrol/src/methods/optional-list/index.js
@@ -1,0 +1,3 @@
+// Importing methods from respective files and exporting them
+
+export { default as addToListMethod } from "./add-to-list";
